+++ google2/CHANGELOG.md
@@ -19,7 +19,7 @@ Added
 Changed
 - Update Google Cloud Storage client library to 1.77.0
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.4-TRAVIS-REPLACE-ME"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.4-74860a5"`
 
 ## 0.3