+++ google2/CHANGELOG.md
@@ -8,7 +8,7 @@ Changed
 - Use linebacker for blocking execution context
 - Moved `org.broadinstitute.dsde.workbench.google.GoogleKmsService` to `org.broadinstitute.dsde.workbench.google2.GoogleKmsService`
 - Add optional generation parameter to `removeObject`
-- Support adding bucket labels in createBucket
+- Add insertBucket, which supports adding bucket labels
 
 SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.5-TRAVIS-REPLACE-ME"`